+++ io-classes/io-classes/Control/Monad/Class/MonadEventlog.hs
@@ -2,7 +2,7 @@ module Control.Monad.Class.MonadEventlog (MonadEventlog (..)) where
 
 import Control.Monad.Reader
 
-import Debug.Trace qualified as IO (traceEventIO, traceMarkerIO)
+import Debug.Trace qualified as IO
 
 class Monad m => MonadEventlog m where
 
@@ -17,13 +17,18 @@ class Monad m => MonadEventlog m where
   -- profiling tools to help you keep clear which marker is which.
   traceMarkerIO :: String -> m ()
 
+  -- | Immediately flush the event log, if enabled.
+  --
+  flushEventLog :: m ()
+
 --
 -- Instances for IO
 --
 
 instance MonadEventlog IO where
-  traceEventIO = IO.traceEventIO
+  traceEventIO  = IO.traceEventIO
   traceMarkerIO = IO.traceMarkerIO
+  flushEventLog = IO.flushEventLog
 
 --
 -- Instance for ReaderT
@@ -32,3 +37,4 @@ instance MonadEventlog IO where
 instance MonadEventlog m => MonadEventlog (ReaderT r m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog